revise table validation, apply check to figures, move code outside the fulltext parser
