+++ grobid-core/src/main/java/org/grobid/core/data/Table.java
@@ -30,7 +30,6 @@ import java.util.List;
 import nu.xom.Attribute;
 import nu.xom.Element;
 import nu.xom.Node;
-import nu.xom.Text;
 
 import static org.grobid.core.document.xml.XmlBuilderUtils.teiElement;
 import static org.grobid.core.document.xml.XmlBuilderUtils.addXmlId;
@@ -44,8 +43,6 @@ public class Table extends Figure {
 	private List<LayoutToken> contentTokens = new ArrayList<>();
 	private List<LayoutToken> fullDescriptionTokens = new ArrayList<>();
 
-    // Contains the raw layoutTokens from the fulltext model
-    private List<LayoutToken> rawLayoutTokens = new ArrayList<>();
 	private boolean goodTable = true;
 
     private StringBuilder note = null;
@@ -65,9 +62,13 @@ public class Table extends Figure {
         note = new StringBuilder();
     }
 
+    public boolean isCompleteForTEI() {
+        return (StringUtils.isNotEmpty(header) && StringUtils.isNotEmpty(caption));
+    }
+
 	@Override
     public String toTEI(GrobidAnalysisConfig config, Document doc, TEIFormatter formatter, List<MarkerType> markerTypes) {
-		if (StringUtils.isEmpty(header) && StringUtils.isEmpty(caption)) {
+		if (!isCompleteForTEI()) {
 			return null;
 		}
 
@@ -107,7 +108,7 @@ public class Table extends Figure {
                 addXmlId(desc, "_" + divID);
             }
 
-            if ( (labeledCaption != null) && (labeledCaption.length() > 0) ) {
+            if (StringUtils.isNotBlank(labeledCaption)) {
                 TaggingTokenClusteror clusteror = new TaggingTokenClusteror(GrobidModels.FULLTEXT, labeledCaption, captionLayoutTokens);
                 List<TaggingTokenCluster> clusters = clusteror.cluster();                
                 for (TaggingTokenCluster cluster : clusters) {
@@ -172,7 +173,7 @@ public class Table extends Figure {
 		}
 
         Element noteNode = null;
-        if (note != null && note.toString().trim().length()>0) {
+        if (StringUtils.isNotBlank(note)) {
 
             noteNode = XmlBuilderUtils.teiElement("note");
             if (config.isGenerateTeiIds()) {
@@ -180,7 +181,7 @@ public class Table extends Figure {
                 addXmlId(noteNode, "_" + divID);
             }
 
-            if ( (labeledNote != null) && (labeledNote.length() > 0) ) {
+            if (StringUtils.isNotBlank(labeledNote)) {
                 TaggingTokenClusteror clusteror = new TaggingTokenClusteror(GrobidModels.FULLTEXT, labeledNote, noteLayoutTokens);
                 List<TaggingTokenCluster> clusters = clusteror.cluster();                
                 for (TaggingTokenCluster cluster : clusters) {
@@ -349,9 +350,14 @@ public class Table extends Figure {
         return this.labeledNote;
     }
 
-	private boolean validateTable() {
+    /** Check if the table:
+     * - has label, header and content
+     * - header starts with "tab"
+     * - label can be parsed
+     */
+	public boolean validateTable() {
 		CntManager cnt = Engine.getCntManager();
-		if (StringUtils.isEmpty(label) || StringUtils.isEmpty(header) || StringUtils.isEmpty(content)) {
+		if (StringUtils.isAnyBlank(label, header, content)) {
 			cnt.i(TableRejectionCounters.EMPTY_LABEL_OR_HEADER_OR_CONTENT);
 			return false;
 		}
@@ -362,7 +368,8 @@ public class Table extends Figure {
 			cnt.i(TableRejectionCounters.CANNOT_PARSE_LABEL_TO_INT);
 			return false;
 		}
-		if (!getHeader().toLowerCase().startsWith("table")) {
+        // tab covers: table, tabelle, tableu, tabella, etc.
+		if (!StringUtils.startsWithIgnoreCase(getHeader(), "tab")) {
 			cnt.i(TableRejectionCounters.HEADER_NOT_STARTS_WITH_TABLE_WORD);
 			return false;
 		}
@@ -427,11 +434,4 @@ public class Table extends Figure {
         return "tab_" + this.id;
     }
 
-    public List<LayoutToken> getRawLayoutTokens() {
-        return rawLayoutTokens;
-    }
-
-    public void setRawLayoutTokens(List<LayoutToken> rawLayoutTokens) {
-        this.rawLayoutTokens = rawLayoutTokens;
-    }
 }
